Calculos de IMSS Patronales

+++ database/migrations/2020_08_05_141624_create_empresas_table.php
@@ -37,7 +37,7 @@ class CreateEmpresasTable extends Migration
             $table->string('tipoPeriodo',10);
             $table->date('inicioPeriodo');
             $table->string('region',50);
-            $table->float('primaRiesgo');
+            $table->double('primaRiesgo',8,4);
             $table->double('porcentajeAhorro');
             $table->timestamps();
         });
